# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation for Dihauti Residence",
        "description": "description",
        "contact": {
            "email": "dev.arven@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.dihauti.ru/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get all posts with pagination",
                "description": "Returns all posts with pagination",
                "operationId": "getPostsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blog posts retrieved successfully"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create yet another post in database",
                "operationId": "postCreate",
                "responses": {
                    "201": {
                        "description": "Blog post created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        },
        "/posts/{slug}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get post by slug",
                "description": "Returns post by slug",
                "operationId": "getPostBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Post slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post retrieved successfully"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Update post in database",
                "operationId": "postUpdate",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Post slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Post updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete post in database",
                "operationId": "postDelete",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Post slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Post deleted successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        },
        "/sortedPost": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get sorting posts",
                "description": "Returns sorting posts",
                "operationId": "getSortedBlogPostList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/posts/{blogPost}/like": {
            "post": {
                "tags": [
                    "Like"
                ],
                "summary": "Like post",
                "operationId": "postLike",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Post slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Liked this post successfully"
                    },
                    "202": {
                        "description": "Unliked this post successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/login",
                        "refreshUrl": "/logout",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    }
}